Improve UI for form errors (#1112)
* Add individual field errors in admin/settings route

* Fix i18n for editing in admin settings

* Remove unused code in /organizations

* Sort data for MultiSelectBox on server

* Add individual field errors in /users/invite

* Add individual field errors in /projects/new

* Slightly better errors for project import

* Include zerodevx/svelte-toast

* Add simpler toast cases from S1

* Fix toast stacking

* Add comment noting remaining locations for toasts

* Require WorkflowDefinition.Name

* Remove unnecessary error returns

Returning form.errors is redundant. Already handled by returning form

* Use consistent validation for org settings

* Add a few more shared schemas for consistency

* Use semantic input types

* Consistent form structure and style

* Use textarea vs input consistently, apply height

* Remove unnecessary id attribute from inputs

* Remove unnecessary logs

* Fix names on org infrastructure page

* Require url/token to be provided when not using default build engine

* Automatic updates of stores/prodDefs in org settings

* Toasts in org settings

* Correct toasts and submission behavior in /users

* Add empty pages for redirects in /users

* Add extra validation to GroupMems/UserRoles

* Improve server-side performance for user profile settings

* Improve project validation performance

Prisma does not currently have an `exists` function. For performance reasons, `findFirst` is recommended over `count` for approximating `exists`

* Prevent TypeaheadInput from submitting a form

The only time submission is desired is when searching by language code in project directory. Submission is handled by the onLanguageCodeSelected event

* Fix timezone selection

* Change user profile form validation

* Consistent casing of Url in form fields

* Consistent return across toggle for org functions

* Remove extraneous console logs

* Consolidate shared valibot schemas

* Fix error return type for addIssue

* Add clarifying comment for UserRoles task enqueuing

* Fix escaping for phone regex, allow country code

+++ portal/common/databaseProxy/OrganizationStores.ts
@@ -1,29 +1,25 @@
 import prisma from '../prisma.js';
 
-export async function updateOrganizationStores(orgId: number, stores: number[]) {
-  const old = (
-    await prisma.organizationStores.findMany({
-      where: {
-        OrganizationId: orgId
-      }
-    })
-  ).map((x) => x.StoreId);
-  const newEntries = stores.filter((store) => !old.includes(store));
-  const removeEntries = old.filter((store) => !stores.includes(store));
-  await prisma.$transaction([
-    prisma.organizationStores.deleteMany({
-      where: {
-        OrganizationId: orgId,
-        StoreId: {
-          in: removeEntries
-        }
-      }
-    }),
-    prisma.organizationStores.createMany({
-      data: newEntries.map((store) => ({
-        OrganizationId: orgId,
-        StoreId: store
-      }))
-    })
-  ]);
+export async function toggleForOrg(
+  OrganizationId: number,
+  StoreId: number,
+  enabled: boolean
+) {
+  if (enabled) {
+    // ISSUE: #1102 this extra check would be unneccessary if we could switch to composite primary keys
+    const existing = await prisma.organizationStores.findFirst({
+      where: { OrganizationId, StoreId },
+      select: { Id: true }
+    });
+    if (!existing) {
+      await prisma.organizationStores.create({
+        data: { OrganizationId, StoreId }
+      });
+    }
+  } else {
+    await prisma.organizationStores.deleteMany({
+      where: { OrganizationId, StoreId }
+    });
+  }
+  return true;
 }
